requirements.txt
openai
python-dotenv
pygame
requests
-e git+https://github.com/openai/swarm.git#egg=swarm
